Working on the CustomizeOrderGui.
Plugin works at this point (it appears)

Progress update.

Still need to work on CustomizeOrderGui... its not done.

Also worked on the "registering worlds" solution. Made some code in CommandReloadWorlds much more readable. Also fixed WorldTPWorldGroup so that it only requires the world name to initialize and not the item and pos.