Changed the way WorldTPWorldSpawnPoint's work... made it make more sense.
Changed the DimensionsSelectionGui so that the resume item only shows if you're not going to the same world that you are already in.

Changed the WorldSelectionGui so that if you click on a world that has a WorldTPWorldSpawnPoint, then you will be teleported directly to that point. Once inside the world, you will be able to access the other dimensions for that world.